FUSETOOLS2-1439 - provide basic test for Yaml route
- the basic test covering the most common use case is now launched with
both Java and Yaml DSL
- not provided a yaml version for other tests, there shouldn't be
something specific to it for now. It would be nice in the future but I
think it already gives enough confidence for now having "just" the most
common scenario covered.
